+++ src/backend/resolvers/queries/index.js
@@ -11,7 +11,6 @@ const collections = require('./collections')
 const newTreatmentPlans = require('./newTreatmentPlans')
 const emailUsers = require('./emailUsers')
 const testEmailGroups = require('./testEmailGroups')
-const queryToolAccounts = require('./queryTool')
 
 module.exports = {
   collections,
@@ -27,5 +26,4 @@ module.exports = {
   emailUsers,
   ...organizations,
   testEmailGroups,
-  queryToolAccounts,
 }